chore: address PR review comments
