+++ dashboard/components/export-csv/ExportCSV.tsx
@@ -1,6 +1,8 @@
+import { useRouter } from 'next/router';
+import { useState } from 'react';
+import settingsService from '../../services/settingsService';
 import { ToastProps } from '../toast/hooks/useToast';
 import ExportCSVButton from './ExportCSVButton';
-import useExportCSV from './useExportCSV';
 
 type ExportCSVProps = {
   displayInTable?: boolean;
@@ -8,12 +10,24 @@ type ExportCSVProps = {
 };
 
 function ExportCSV({ displayInTable = false, setToast }: ExportCSVProps) {
-  const { id, isFilteredList, loading, exportCSV } = useExportCSV({ setToast });
+  const router = useRouter();
+
+  function exportCSV(id?: string) {
+    settingsService.exportCSV(id);
+    setToast({
+      hasError: false,
+      title: 'CSV exported',
+      message: 'The download of the CSV file should begin shortly.'
+    });
+  }
+
+  const isFilteredList =
+    Object.keys(router.query).length > 0 && !router.query.view;
+  const id = router.query.view ? router.query.view.toString() : undefined;
 
   return (
     <ExportCSVButton
       id={id}
-      loading={loading}
       disabled={isFilteredList}
       displayInTable={displayInTable}
       exportCSV={exportCSV}
